Make the parity harness total, and fix two percentage fields that were Int
The parity test could be thrown out of by two things it did not handle: a
model's own validation error, which carries no coding path, and a rejection at
a path it could not strip. Both are now reported with the reason rather than
escaping the test, and an array element the hand type rejects no longer
registers as a count difference against the generated type.

It also surfaced a hand-type defect worth fixing at once: two fields the spec
declares as `number` were `Int` — Subscription.TransferData.amountPercent and
Coupon.percentOff. A 12.5% coupon or an 87.5% transfer split failed to decode
outright. Both are now Decimal, with a decoding test.

+++ Sources/Stripe/Models/Billing/Stripe.Billing.Subscription.swift
@@ -564,7 +564,7 @@ extension Stripe.Billing.Subscription.Trial {
 extension Stripe.Billing.Subscription {
     public struct TransferData: Codable, Hashable, Sendable {
         /// Transfer percentage
-        public var amountPercent: Int?
+        public var amountPercent: Decimal?
         /// Destination account
         @ExpandableOf<Stripe.Connect.Account> public var destination: Stripe.Connect.Account.ID?
 
@@ -574,7 +574,7 @@ extension Stripe.Billing.Subscription {
         }
 
         public init(
-            amountPercent: Int? = nil,
+            amountPercent: Decimal? = nil,
             destination: Stripe.Connect.Account.ID? = nil
         ) {
             self.amountPercent = amountPercent
